add enemy chase effect and make enemy can not pick up potions

+++ Source/Aura/Private/AuraAbilityTypes.cpp
@@ -2,7 +2,7 @@
 
 bool FAuraGameplayEffectContext::NetSerialize(FArchive& Ar, class UPackageMap* Map, bool& bOutSuccess)
 {
-	uint32 RepBits = 0;
+	uint64 RepBits = 0;
 	if (Ar.IsSaving())
 	{
 		if (bReplicateInstigator && Instigator.IsValid())
